Cleanup of QML windowing and message boxes
