+++ interface/resources/qml/AddressBarDialog.qml
@@ -12,37 +12,37 @@ import Hifi 1.0
 import QtQuick 2.4
 import "controls"
 import "styles"
+import "windows"
 
-DialogContainer {
+Window {
     id: root
     HifiConstants { id: hifi }
-    z: 1000
-
+    anchors.centerIn: parent
     objectName: "AddressBarDialog"
 
-    property bool destroyOnInvisible: false
-    property real scale: 1.25  // Make this dialog a little larger than normal
+    destroyOnInvisible: false
+    resizable: false
+    scale: 1.25  // Make this dialog a little larger than normal
 
-    implicitWidth: addressBarDialog.implicitWidth
-    implicitHeight: addressBarDialog.implicitHeight
-
-    x: parent ? parent.width / 2 - width / 2 : 0
-    y: parent ? parent.height / 2 - height / 2 : 0
-    property int maximumX: parent ? parent.width - width : 0
-    property int maximumY: parent ? parent.height - height : 0
-
-
-    Rectangle {
-        id: dragRegion
-        visible: dragMouseArea.containsMouse
-    } 
+    width: addressBarDialog.implicitWidth
+    height: addressBarDialog.implicitHeight
 
     AddressBarDialog {
         id: addressBarDialog
-        z: dragRegion.z + 1
         implicitWidth: backgroundImage.width
         implicitHeight: backgroundImage.height
 
+        Timer {
+            running: root.visible
+            interval: 500
+            repeat: true
+            onTriggered: {
+                if (root.enabled && !addressLine.activeFocus) {
+                    addressLine.forceActiveFocus()
+                }
+            }
+        }
+        
         Image {
             id: backgroundImage
 
@@ -52,45 +52,9 @@ DialogContainer {
             property int inputAreaHeight: 56.0 * root.scale  // Height of the background's input area
             property int inputAreaStep: (height - inputAreaHeight) / 2
 
-            MouseArea {
-                id: dragMouseArea
-                // Drag the icon
-                width: parent.height
-                height: parent.height
-                x: 0
-                y: 0
-                hoverEnabled: true
-                drag {
-                    target: root
-                    minimumX: -parent.inputAreaStep
-                    minimumY: -parent.inputAreaStep
-                    maximumX: root.parent ? root.maximumX : 0
-                    maximumY: root.parent ? root.maximumY + parent.inputAreaStep : 0
-                }
-            }
-            
-            
-
-            MouseArea {
-                // Drag the input rectangle
-                width: parent.width - parent.height
-                height: parent.inputAreaHeight
-                x: parent.height
-                y: parent.inputAreaStep
-                drag {
-                    target: root
-                    minimumX: -parent.inputAreaStep
-                    minimumY: -parent.inputAreaStep
-                    maximumX: root.parent ? root.maximumX : 0
-                    maximumY: root.parent ? root.maximumY + parent.inputAreaStep : 0
-                }
-            }
-
             Image {
                 id: backArrow
-
                 source: addressBarDialog.backEnabled ? "../images/left-arrow.svg" : "../images/left-arrow-disabled.svg"
-                
                 anchors {
                     fill: parent
                     leftMargin: parent.height + hifi.layout.spacing + 6
@@ -110,9 +74,7 @@ DialogContainer {
             
             Image {
                 id: forwardArrow
-
                 source: addressBarDialog.forwardEnabled ? "../images/right-arrow.svg" : "../images/right-arrow-disabled.svg"
-                
                 anchors {
                     fill: parent
                     leftMargin: parent.height + hifi.layout.spacing * 9
@@ -130,6 +92,7 @@ DialogContainer {
                 }
             }
 
+            // FIXME replace with TextField
             TextInput {
                 id: addressLine
 
@@ -139,61 +102,35 @@ DialogContainer {
                     rightMargin: hifi.layout.spacing * 2
                     topMargin: parent.inputAreaStep + hifi.layout.spacing
                     bottomMargin: parent.inputAreaStep + hifi.layout.spacing
-
                 }
-
                 font.pixelSize: hifi.fonts.pixelSize * root.scale * 0.75
-
                 helperText: "Go to: place, @user, /path, network address"
-
-                onAccepted: {
-                    event.accepted = true  // Generates erroneous error in program log, "ReferenceError: event is not defined".
-                    addressBarDialog.loadAddress(addressLine.text)
-                }
             }
 
-        }
-    }
-
-    onEnabledChanged: {
-        if (enabled) {
-            addressLine.forceActiveFocus()
-        }
-    }
-    
-    Timer {
-        running: root.enabled
-        interval: 500
-        repeat: true
-        onTriggered: {
-            if (root.enabled && !addressLine.activeFocus) {
-                addressLine.forceActiveFocus()
-            }
         }
     }
 
     onVisibleChanged: {
-        if (!visible) {
+        if (visible) {
+            addressLine.forceActiveFocus()
+        } else {
             addressLine.text = ""
         }
     }
-
+    
     function toggleOrGo() {
-        if (addressLine.text == "") {
-            enabled = false
-        } else {
+        if (addressLine.text != "") {
             addressBarDialog.loadAddress(addressLine.text)
         }
+        root.close();
     }
 
     Keys.onPressed: {
         switch (event.key) {
             case Qt.Key_Escape:
             case Qt.Key_Back:
-                if (enabled) {
-                    enabled = false
-                    event.accepted = true
-                }
+                root.close()
+                event.accepted = true
                 break
             case Qt.Key_Enter:
             case Qt.Key_Return:
