+++ interface/resources/qml/InfoView.qml
@@ -1,11 +1,10 @@
+import QtQuick 2.5
 import Hifi 1.0 as Hifi
-import QtQuick 2.3
-import QtQuick.Controls 1.2
-import QtQuick.Controls.Styles 1.3
-import QtWebEngine 1.1
-import "controls"
 
-VrDialog {
+import "controls"
+import "windows" as Windows
+
+Windows.Window {
     id: root
     width: 800
     height: 800
@@ -13,12 +12,10 @@ VrDialog {
     
     Hifi.InfoView {
         id: infoView
-        // Fille the client area
+        // Fill the client area
         anchors.fill: parent
-        anchors.margins: parent.margins
-        anchors.topMargin: parent.topMargin
  
-        WebEngineView {
+        WebView {
             id: webview
             objectName: "WebView"
             anchors.fill: parent
