--- interface/src/ui/RecorderDialog.cpp
@@ -1,22 +1,0 @@
-//
-//  Created by Bradley Austin Davis on 2015/11/14
-//  Copyright 2015 High Fidelity, Inc.
-//
-//  Distributed under the Apache License, Version 2.0.
-//  See the accompanying file LICENSE or http://www.apache.org/licenses/LICENSE-2.0.html
-//
-
-#include "RecorderDialog.h"
-
-#include <QMessageBox>
-
-#include "DependencyManager.h"
-
-HIFI_QML_DEF(RecorderDialog)
-
-RecorderDialog::RecorderDialog(QQuickItem* parent) : OffscreenQmlDialog(parent) {
-}
-
-void RecorderDialog::hide() {
-    ((QQuickItem*)parent())->setEnabled(false);
-}
