+++ tests/ui/qml/Palettes.qml
