+++ tests/ui/qml/TestDialog.qml
