+++ tests/ui/qml/controlDemo/ButtonPage.qml
