+++ tests/ui/qml/controlDemo/InputPage.qml
