+++ tests/ui/qml/controlDemo/ProgressPage.qml
