+++ tests/ui/qml/controlDemo/main.qml
